Move other PSF functions into module
