+++ src/PSF.jl
@@ -12,16 +12,73 @@ using Celeste.SensitiveFloats.SensitiveFloat
 using Celeste.SensitiveFloats.clear!
 
 import Optim
+import WCS
 
 export evaluate_psf_fit, psf_params_to_array, psf_array_to_params,
        get_psf_transform, initialize_psf_params, transform_psf_params!,
        unwrap_psf_params, wrap_psf_params,
        unconstrain_psf_params, constrain_psf_params,
        transform_psf_sensitive_float!,
-       PsfOptimizer, fit_raw_psf_for_celeste
+       PsfOptimizer, fit_raw_psf_for_celeste,
+       get_psf_at_point, get_source_psf
 
  # Only include until this is merged with Optim.jl.
- include("newton_trust_region.jl")
+include("newton_trust_region.jl")
+
+"""
+Return an image of a Celeste GMM PSF evaluated at rows, cols.
+
+Args:
+- psf_array: The PSF to be evaluated as an array of PsfComponent
+- rows: The rows in the image (usually in pixel coordinates)
+- cols: The column in the image (usually in pixel coordinates)
+
+Returns:
+ - The PSF values at rows and cols.  The default size is the same as
+   that returned by get_psf_at_point applied to FITS header values.
+
+Note that the point in the image at which the PSF is evaluated --
+that is, the center of the image returned by this function -- is
+already implicit in the value of psf_array.
+"""
+function get_psf_at_point(psf_array::Array{PsfComponent, 1};
+                         rows=collect(-25:25), cols=collect(-25:25))
+
+  function get_psf_value(psf::PsfComponent, row::Float64, col::Float64)
+     x = Float64[row, col] - psf.xiBar
+     exp_term = exp(-0.5 * x' * psf.tauBarInv * x - 0.5 * psf.tauBarLd)
+     (psf.alphaBar * exp_term / (2 * pi))[1]
+  end
+
+  Float64[
+   sum([ get_psf_value(psf, float(row), float(col)) for psf in psf_array ])
+     for row in rows, col in cols ]
+end
+
+
+"""
+Get the PSF located at a particular world location in an image.
+
+Args:
+ - world_loc: A location in world coordinates.
+ - img: An Image
+
+Returns:
+ - An array of PsfComponent objects that represents the PSF as a mixture
+   of Gaussians.
+"""
+function get_source_psf(world_loc::Vector{Float64}, img::Image)
+  # Some stamps or simulated data have no raw psf information.  In that case,
+  # just use the psf from the image.
+  if size(img.raw_psf_comp.rrows) == (0, 0)
+    return img.psf
+  else
+    pixel_loc = WCS.world_to_pix(img.wcs, world_loc)
+    psfstamp = img.raw_psf_comp(pixel_loc[1], pixel_loc[2])
+    return PSF.fit_raw_psf_for_celeste(psfstamp)
+  end
+end
+
 
 function initialize_psf_params(K::Int; for_test::Bool=false)
   psf_params = Array(Vector{Float64}, K)
