+++ test/test_images.jl
@@ -1,11 +1,12 @@
 using Base.Test
 using DataFrames
 
-import SloanDigitalSkySurvey.PSF
+#import SloanDigitalSkySurvey.PSF
 import WCS
 
 using Celeste: Types, SampleData
-import Celeste: ModelInit, SkyImages, ElboDeriv, Synthetic, SDSSIO
+#import SloanDigitalSkySurvey
+import Celeste: ModelInit, SkyImages, ElboDeriv, Synthetic, SDSSIO, PSF
 import Celeste.ModelInit: patch_ctrs_pix, patch_radii_pix
 
 println("Running SkyImages tests.")
@@ -70,7 +71,7 @@ function test_blob()
   original_psf_celeste = PSF.fit_raw_psf_for_celeste(original_psf_val);
   fit_original_psf_val = PSF.get_psf_at_point(original_psf_celeste);
 
-  obj_psf = SkyImages.get_source_psf(mp_obj.vp[1][ids.u], img);
+  obj_psf = PSF.get_source_psf(mp_obj.vp[1][ids.u], img);
   obj_psf_val = PSF.get_psf_at_point(obj_psf);
 
   # The fits should match exactly.
@@ -97,7 +98,7 @@ function test_stamp_get_object_psf()
   original_psf_val = PSF.get_psf_at_point(img.psf);
 
   obj_psf_val =
-    PSF.get_psf_at_point(SkyImages.get_source_psf(stamp_mp.vp[1][ids.u], img))
+    PSF.get_psf_at_point(PSF.get_source_psf(stamp_mp.vp[1][ids.u], img))
   @test_approx_eq_eps(obj_psf_val, original_psf_val, 1e-6)
 end
